Reveal the final artist round; fix guess-distribution counts
Two bugs reported from play:

Skipping the tenth song never showed its answer. On the last round the client
set status straight to 'completed', so the reveal step every other round gets
was skipped entirely and the summary replaced it — the one song whose answer a
player never learned was the one they had just given up on. The reveal now
shows for the final round too, with the button reading "See results →" instead
of "Next song →", and pressing it goes to the summary rather than refetching a
finished challenge.

The guess-distribution legend read "All players (01800000)" and "avg 0.0s".
Postgres COUNT(*) is bigint, which postgres-js returns as a *string* to avoid
precision loss; the row type was a hand-written assertion claiming number, so
nothing caught it. The client then summed buckets with `+`, concatenating
instead of adding, and divided the average by that. Cast to ::int in SQL.

Both are covered by tests that were confirmed to fail against the old code:
the distribution test asserts the runtime type (a type assertion cannot), and
the hook tests assert the final round lands on 'round-ended' with the song
revealed, then reaches the summary only on the next press.

+++ apps/server/src/services/artistChallengeService.ts
@@ -555,8 +555,13 @@ export async function getArtistGuessDistribution(
 ): Promise<GuessDistributionBucket[]> {
   const deezerArtistId = String(artistId);
 
+  // The `::int` casts are load-bearing. Postgres COUNT(*) is bigint, which postgres-js returns
+  // as a *string* to avoid precision loss, and the row type assertion below cannot catch that.
+  // The client then summed the buckets with `+`, concatenating instead of adding — the legend
+  // read "All players (01800000)" and the average divided by that, showing "avg 0.0s".
+
   const allRows = (await db.execute(sql`
-    SELECT g.snippet_stage_seconds as "snippetStageSeconds", COUNT(*) as count
+    SELECT g.snippet_stage_seconds as "snippetStageSeconds", COUNT(*)::int as count
     FROM artist_round_guesses g
     JOIN artist_session_results s ON s.id = g.session_id
     JOIN artist_challenges c ON c.id = s.challenge_id
@@ -567,7 +572,7 @@ export async function getArtistGuessDistribution(
   `)) as unknown as { snippetStageSeconds: number; count: number }[];
 
   const myRows = (await db.execute(sql`
-    SELECT g.snippet_stage_seconds as "snippetStageSeconds", COUNT(*) as count
+    SELECT g.snippet_stage_seconds as "snippetStageSeconds", COUNT(*)::int as count
     FROM artist_round_guesses g
     JOIN artist_session_results s ON s.id = g.session_id
     JOIN artist_challenges c ON c.id = s.challenge_id
